transactions_backend/requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.13.1
install==1.3.5
psycopg2-binary==2.9.3
pydantic==1.9.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
